Add solution for LeetCode - Average Salary Excluding the Minimum and Maximum Salary

+++ LeetCode/Easy/1491 Average Salary Excluding the Minimum and Maximum Salary/Solution.cpp
@@ -7,7 +7,7 @@ public:
             sum=sum+salary[i];
         }
         int up=sum-salary[0]-salary[salary.size()-1];
-        int avg=up/(salary.size()-2);
+        double avg=(double)up/(salary.size()-2);
         return avg;
     }
 };
